Server rafactor (#22)
* feat: server refactor

* fix: server client connectino

* fix: got client/server working

* fix:a dded comments

* feat: started tracker

* fix: bug fixes

* feat: started development on bittorrent

* feat: Added torrent file parser (#29)

* initialice torrent


* continue init torrent struct

* parsed single file torrents through getters

* torrent_file: (get_info_hash is not working) :(

* fix info_hash :)


* fix: dynamic port search

* misc

* First protocol Messages

* HTTPTracker send_handshake

* WIP protocol Messages

* feat: Urlencoder (#30)

* inicial urlencoder


* add decode and encode functionality


* added README to urlencode


* fix: fixed hash encoding

* fix: fixed request

* feat: connected to first peer

* fix: misc

* fix: misc

* fix: implemented messages

* fix: changed texts

* fix: message modification

* fix: IT WORKS

* Main .torrent pathname

* FIX: set env vars

* Cargo fmt

* Remove unwraps

* fix: downloads first piece

* fix: removed piece file

* fix: refactorer peer

* fix: fixed error handling

* fix: fixed error handling

* fix: fixed some clippy warnings

* fix: removed all clippy warns

* fix: run format

* fix: fixed tests

+++ src/bencoder/common.rs
@@ -9,14 +9,14 @@ pub const END: u8 = b'e';
 
 #[derive(Debug)]
 pub enum Types {
-    Integer(i32),
+    Integer(i64),
     String(Vec<u8>),
     Dictionary(HashMap<Vec<u8>, Types>),
     List(LinkedList<Types>),
 }
 
 impl Types {
-    pub fn get_integrer(&self) -> i32 {
+    pub fn get_integrer(&self) -> i64 {
         match &self {
             Types::Integer(int) => *int,
             _ => unreachable!(),
